Moved the summary subcommand into its own file. Also removed the commented code.
